+++ posts/views.py
@@ -26,14 +26,16 @@ class ShowPost(CreateView,ListView):
 	'''def get_queryset(self):
         self.post = get_object_or_404(Post, name=self.args[0])
         return Post.objects.filter(post=self.post)'''
-	def get_context_data(self, **kwargs):
-    # Call the base implementation first to get a context
-		context = super(ShowPost, self).get_context_data(**kwargs)
-    # Add in the publisher
-		context['posts'] = Post.objects.all().order_by("-time")
-		context['bulletins'] = Bulletin.objects.all().order_by("-time")
-		return context
+
+	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
+		searchPost = Post.objects.all().order_by("-time")
+
+		return render(request, self.template_name, {'posts':searchPost,'bulletins' : Bulletin.objects.all().order_by("-time")})	
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newdoc = Post()
 		newdoc.title = request.POST['title']
 		newdoc.context = request.POST['context']
@@ -46,6 +48,8 @@ class PostDetail(DetailView):
 	model = Post
 	template_name = 'post_detail.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newdoc = Post.objects.filter(author = request.user).order_by("-time")
 		searchPost = Post.objects.get(pk=kwargs['pk'])
 		allmsg = Message.objects.filter(belong_post=searchPost).order_by("-time")
@@ -53,6 +57,8 @@ class PostDetail(DetailView):
 		return render(request, self.template_name, {'post':searchPost, 'allmsg': allmsg, 'bulletins' : Bulletin.objects.all().order_by("-time")})	
 
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newmsg = Message()
 		newmsg.context = request.POST['context']
 		newmsg.author = MyProfile.objects.get(user=request.user)
@@ -70,9 +76,13 @@ class PastPostDiscuss(ListView):
 	model = Post
 	template_name = 'pastpost_discuss.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newdoc = Post.objects.filter(author = request.user).order_by("-time")
 		return render(request, self.template_name, {'posts': newdoc, 'bulletins' : Bulletin.objects.all().order_by("-time")})		
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		msgId = request.POST.get('post', None)
 		msgToDel = get_object_or_404(Post, pk = msgId)
 		msgToDel.delete()
@@ -81,9 +91,13 @@ class PastPostDiscuss(ListView):
 class PostEdit(ListView):
 	template_name = 'post_edit.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		msg = get_object_or_404(Post, pk = kwargs['pk'])
 		return render(request, self.template_name, {'post': msg, 'bulletins' : Bulletin.objects.all().order_by("-time")})		
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		msgId = request.POST.get('post', None)
 		msgToEdit = get_object_or_404(Post, pk = msgId)
 		#msgToEdit.update(title = request.POST['title'], context = request.POST['context'], tag1 = request.POST['tag1'])
@@ -96,9 +110,13 @@ class PostEdit(ListView):
 class PostBulletin(ListView):
 	template_name = 'post_bulletin.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		bulletin = Bulletin.objects.filter(author = request.user).order_by("-time")
 		return render(request, self.template_name, {'bulletins': bulletin})	
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newBulletin = Bulletin()
 		newBulletin.title = request.POST['title']
 		newBulletin.context = request.POST['context']
@@ -110,6 +128,8 @@ class PostBulletin(ListView):
 class ViewBulletin(ListView):
 	template_name = 'view_bulletin.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		bulletin = Bulletin.objects.filter(author = request.user).order_by("-time")
 		data = get_object_or_404(Bulletin, pk = kwargs['pk'])
 		return render(request, self.template_name, {'bulletins': bulletin,'data' : data})
@@ -117,10 +137,14 @@ class ViewBulletin(ListView):
 class EditBulletin(ListView):
 	template_name = 'edit_bulletin.html'
 	def get(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		bulletin = Bulletin.objects.filter(author = request.user).order_by("-time")
 		data = get_object_or_404(Bulletin, pk = kwargs['pk'])
 		return render(request, self.template_name, {'bulletins': bulletin,'data' : data})
 	def post(self, request, *args, **kwargs):
+		if not request.user.has_perm('accounts.view_profile'):
+			return render(request, '401.html')
 		newBulletin = get_object_or_404(Bulletin, pk = kwargs['pk'])
 		newBulletin.title = request.POST['title']
 		newBulletin.context = request.POST['context']
